pyproject.toml
[project]
name = "jim-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markdownify>=1.2.2",
    "playwright>=1.58.0",
    "tinydb>=4.8.0",
    "pyyaml>=6.0",
    "fastmcp>=2.0.0",
]
